requirements.txt
python-dotenv==0.21.0
jupyter==1.0.0
matplotlib==3.5.2
numpy==1.23.3
pandas==1.4.4
scikit-learn==1.1.2
tensorflow==2.8.2
tqdm==4.64.1
transformers==4.18.0
tweepy==4.10.1
boto3==1.26.15
openpyxl==3.0.10